[DPE-9158] Limit repo listing to find the timelines (#1403)
* Limit bucket listing to find the timelines

* Add ceph pitr test

* Switch back to recurse

* Refactor tests

* Fix imports

* Fix tests

* Reduce boto logs

* Typo
