+++ tests/spread/test_backups_pitr_ceph.py/task.yaml
@@ -1,9 +1,7 @@
-summary: test_backups_pitr_aws.py
+summary: test_backups_pitr_ceph.py
 environment:
-  TEST_MODULE: test_backups_pitr_aws.py
+  TEST_MODULE: test_backups_pitr_ceph.py
 execute: |
   tox run -e integration -- "tests/integration/$TEST_MODULE" --model testing --alluredir="$SPREAD_TASK/allure-results"
 artifacts:
   - allure-results
-backends:
-  - -lxd-vm  # Requires CI secrets
