+++ tests/spread/test_backups_pitr_gcp.py/task.yaml
@@ -7,5 +7,3 @@ artifacts:
   - allure-results
 backends:
   - -lxd-vm  # Requires CI secrets
-systems:
-  - -ubuntu-24.04-arm
